requirements.txt
Django>=4.2
psycopg2-binary
pgvector
google-generativeai
beautifulsoup4
requests
python-dotenv
apscheduler
lxml
djangorestframework
sentence-transformers
numpy
pypdf
python-docx
feedparser
django-apscheduler
icalendar
